Locating retry logic to Normalization Processor
