+++ src/console/Normalize/Processor.cs
@@ -1,7 +1,21 @@
-﻿using EnsureThat;
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Azure;
+using Azure.Identity;
+using EnsureThat;
 using Microsoft.Azure.EventHubs;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Options;
+using Microsoft.Health.Common.Telemetry;
 using Microsoft.Health.Events.EventConsumers;
 using Microsoft.Health.Events.Model;
 using Microsoft.Health.Fhir.Ingest.Console.Template;
@@ -10,10 +24,7 @@ using Microsoft.Health.Fhir.Ingest.Service;
 using Microsoft.Health.Fhir.Ingest.Telemetry;
 using Microsoft.Health.Fhir.Ingest.Template;
 using Microsoft.Health.Logging.Telemetry;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
+using Polly;
 using static Microsoft.Azure.EventHubs.EventData;
 
 namespace Microsoft.Health.Fhir.Ingest.Console.Normalize
@@ -26,6 +37,7 @@ namespace Microsoft.Health.Fhir.Ingest.Console.Normalize
         private IAsyncCollector<IMeasurement> _collector;
         private IOptions<NormalizationServiceOptions> _normalizationOptions;
         private IEventProcessingMeter _eventProcessingMeter = new EventProcessingMeter();
+        private AsyncPolicy _retryPolicy;
 
         public Processor(
             [Blob("template/%Template:DeviceContent%", FileAccess.Read)] string templateDefinition,
@@ -39,9 +51,17 @@ namespace Microsoft.Health.Fhir.Ingest.Console.Normalize
             _collector = collector;
             _logger = logger;
             _normalizationOptions = options;
+            _retryPolicy = CreateRetryPolicy(logger);
         }
 
         public async Task ConsumeAsync(IEnumerable<IEventMessage> events)
+        {
+            EnsureArg.IsNotNull(events);
+            await _retryPolicy.ExecuteAsync(async () => await ConsumeAsyncImpl(events));
+            
+        }
+
+        private async Task ConsumeAsyncImpl(IEnumerable<IEventMessage> events)
         {
             EnsureArg.IsNotNull(_templateDefinition);
             var templateContent = _templateManager.GetTemplateAsString(_templateDefinition);
@@ -89,6 +109,45 @@ namespace Microsoft.Health.Fhir.Ingest.Console.Normalize
                     IomtMetrics.DeviceIngressSizeBytes(),
                     eventStats.TotalEventsProcessedBytes);
             }
+        }
+
+        private static AsyncPolicy CreateRetryPolicy(ITelemetryLogger logger)
+        {
+            bool ExceptionRetryableFilter(Exception ee)
+            {
+                var type = ee.GetType().ToString();
+                var ToMetric = new Metric(
+                type,
+                new Dictionary<string, object>
+                {
+                        { DimensionNames.Name, type },
+                        { DimensionNames.Category, Category.Errors },
+                        { DimensionNames.ErrorType, ErrorType.DeviceMessageError },
+                        { DimensionNames.ErrorSeverity, ErrorSeverity.Warning },
+                        { DimensionNames.Operation, ConnectorOperation.Normalization},
+                });
+                logger.LogMetric(ToMetric, 1);
+
+                switch (ee)
+                {
+                    case AggregateException ae when ae.InnerExceptions.Any(ExceptionRetryableFilter):
+                    case OperationCanceledException _:
+                    case HttpRequestException _:
+                    case EventHubsException _:
+                    case AuthenticationFailedException _:
+                    case RequestFailedException _:
+                        break;
+                    default:
+                        return false;
+                }
+
+                logger.LogError(new Exception("Encountered retryable exception", ee));
+                return true;
+            }
+
+            return Policy
+                .Handle<Exception>(ExceptionRetryableFilter)
+                .WaitAndRetryForeverAsync(retryCount => TimeSpan.FromSeconds(Math.Min(30, Math.Pow(2, retryCount))));
         }
     }
 }
