+++ src/lib/Microsoft.Health.Events/EventConsumers/Service/EventConsumerService.cs
@@ -6,15 +6,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net.Http;
 using System.Threading.Tasks;
-using Azure;
-using Azure.Identity;
-using Azure.Messaging.EventHubs;
 using EnsureThat;
 using Microsoft.Health.Events.Model;
 using Microsoft.Health.Logging.Telemetry;
-using Polly;
 
 namespace Microsoft.Health.Events.EventConsumers.Service
 {
@@ -23,19 +18,11 @@ namespace Microsoft.Health.Events.EventConsumers.Service
         private readonly IEnumerable<IEventConsumer> _eventConsumers;
         private ITelemetryLogger _logger;
 
-        public EventConsumerService(
-            IEnumerable<IEventConsumer> eventConsumers,
-            ITelemetryLogger logger,
-            bool shouldRetry = true,
-            Action<Exception> exceptionTelemetryProcessor = null)
+        public EventConsumerService(IEnumerable<IEventConsumer> eventConsumers, ITelemetryLogger logger)
         {
             _eventConsumers = EnsureArg.IsNotNull(eventConsumers, nameof(eventConsumers));
             _logger = EnsureArg.IsNotNull(logger, nameof(logger));
-
-            RetryPolicy = CreateRetryPolicy(logger, shouldRetry, exceptionTelemetryProcessor);
         }
-
-        public AsyncPolicy RetryPolicy { get;  }
 
         public Task ConsumeEvent(IEventMessage eventArg)
         {
@@ -50,7 +37,7 @@ namespace Microsoft.Health.Events.EventConsumers.Service
                 {
                     try
                     {
-                        await RetryPolicy.ExecuteAsync(async () => await eventConsumer.ConsumeAsync(events).ConfigureAwait(false));
+                        await eventConsumer.ConsumeAsync(events).ConfigureAwait(false);
                     }
                     catch (Exception e)
                     {
@@ -58,42 +45,6 @@ namespace Microsoft.Health.Events.EventConsumers.Service
                     }
                 }
             }
-        }
-
-        private static AsyncPolicy CreateRetryPolicy(ITelemetryLogger logger, bool shouldRetry, Action<Exception> exceptionTelemetryProcessor)
-        {
-            bool ExceptionRetryableFilter(Exception ee)
-            {
-                if (exceptionTelemetryProcessor != null)
-                {
-                    exceptionTelemetryProcessor.Invoke(ee);
-                }
-
-                if (!shouldRetry)
-                {
-                    return false;
-                }
-
-                switch (ee)
-                {
-                    case AggregateException ae when ae.InnerExceptions.Any(ExceptionRetryableFilter):
-                    case OperationCanceledException _:
-                    case HttpRequestException _:
-                    case EventHubsException _:
-                    case AuthenticationFailedException _:
-                    case RequestFailedException _:
-                        break;
-                    default:
-                        return false;
-                }
-
-                logger.LogError(new Exception("Encountered retryable exception", ee));
-                return true;
-            }
-
-            return Policy
-                .Handle<Exception>(ExceptionRetryableFilter)
-                .WaitAndRetryForeverAsync(retryCount => TimeSpan.FromSeconds(Math.Min(30, Math.Pow(2, retryCount))));
         }
     }
 }
